+++ native/c/zam.h
@@ -71,7 +71,7 @@ void read_dcb(IHADCB *, READ_PL *, char *) ATTRIBUTE(amode31);
 #pragma map(write_output_bpam, "WRTOBPAM")
 #endif
 
-int open_output_bpam(ZDIAG *PTR32, IO_CTRL *PTR32 *PTR32, const char *PTR32) ATTRIBUTE(amode31);
+int open_output_bpam(ZDIAG *PTR32, IO_CTRL * PTR32 * PTR32, const char *PTR32) ATTRIBUTE(amode31);
 int write_output_bpam(ZDIAG *PTR32, IO_CTRL *PTR32, const char *PTR32, int length) ATTRIBUTE(amode31);
 int close_output_bpam(ZDIAG *PTR32, IO_CTRL *PTR32) ATTRIBUTE(amode31);
 
@@ -81,7 +81,7 @@ int close_output_bpam(ZDIAG *PTR32, IO_CTRL *PTR32) ATTRIBUTE(amode31);
 #pragma map(point_input_vsam, "PTNIVSAM")
 #pragma map(close_input_vsam, "CLSIVSAM")
 #endif
-int open_input_vsam(ZDIAG *PTR32 diag, IO_CTRL *PTR32 *PTR32, const char *PTR32) ATTRIBUTE(amode31);
+int open_input_vsam(ZDIAG *PTR32 diag, IO_CTRL * PTR32 * PTR32, const char *PTR32) ATTRIBUTE(amode31);
 int read_input_vsam(ZDIAG *PTR32 diag, IO_CTRL *PTR32) ATTRIBUTE(amode31);
 int point_input_vsam(ZDIAG *PTR32 diag, IO_CTRL *PTR32, TIME_STRUCT *time_struct) ATTRIBUTE(amode31);
 int close_input_vsam(ZDIAG *PTR32, IO_CTRL *PTR32) ATTRIBUTE(amode31);