+++ packages/cli/src/list/List.definition.ts
@@ -12,6 +12,7 @@
 import type { ICommandDefinition } from "@zowe/imperative";
 import { SshSession } from "@zowe/zos-uss-for-zowe-sdk";
 import { Constants } from "../Constants";
+import { ListAliasDefinition } from "./alias/Alias.definition";
 import { ListDataSetDefinition } from "./data-set/DataSet.definition";
 import { ListDataSetMembersDefinition } from "./data-set-members/DataSetMembers.definition";
 import { ListJobsDefinition } from "./jobs/Jobs.definition";
@@ -25,6 +26,7 @@ const ListDefinition: ICommandDefinition = {
     description: "List data sets, data set members, uss files, jobs, spool files",
     type: "group",
     children: [
+        ListAliasDefinition,
         ListDataSetDefinition,
         ListDataSetMembersDefinition,
         ListUssFilesDefinition,