+++ packages/cli/CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to the Client code for "zowex-for-zowe-cli" are documented i
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## Recent Changes
+## Recent Changes 
 
+- Added the `zssh files list alias` command to the CLI. [#1108](https://github.com/zowe/zowex/pull/1108)
 - Added support for SSH agents (`IdentityAgent` in SSH config) with the `--agent` option. [#342](https://github.com/zowe/zowex/issues/342)
+
+## `0.9.0`
+
+- Added the `zssh issue console-command` command. The `--console-name` option is now optional and defaults to the current user ID. [#1112](https://github.com/zowe/zowex/issues/1112)
 
 ## `0.8.0`